refactor: add src entry point (#28)
