+++ docs/README.md
@@ -25,5 +25,6 @@ Reference these documents when:
 ## Related modules
 
 * [`tests/`](../tests/) – Test fixtures and implementations for the rules
+* [`src/`](../src/) – General library code and package entry point
 * [`vscode/custom-rules`](../.vscode/custom-rules/) – Implementation of the custom rules
 * [`README.md`](../README.md) – Project overview
